% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.EOT.R
\name{fast.eot}
\alias{fast.eot}
\title{A parallelized version to calculate  EOTs (Empirical orthogonal teleconnections)}
\usage{
fast.eot(
  x,
  n = 1,
  standardised,
  n_cores = parallel::detectCores() - 1,
  verbose = T
)
}
\arguments{
\item{x}{The spatio-temporal data, organised in a 'RasterBrick'-format.}

\item{n}{The number of EOTs to calculate.}

\item{standardised}{Boolean. Indicates whether the output should be scaled or not.}

\item{n_cores}{Integer. The number of CPU-cores for the function to use.}

\item{verbose}{Boolean. Should intermediate messages be printed to the console (TRUE) or not (FALSE). Default to TRUE.}
}
\value{
A List of the EOTs.
}
\description{
This function builds on the work done for the 'remote' package, but parallelizes the code, in order to reduce computation times.
Given a spatio-temporal field (as a raster/brick file) it calculates a predefined number of EOTs.
}
