% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eot.R
\name{plot.eot}
\alias{plot.eot}
\title{Function to plot the EOT pattern (spatial/temporal)}
\usage{
\method{plot}{eot}(eot.out, n, type = "regression weights")
}
\arguments{
\item{eot.out}{An output object from the \code{fast.EOT} function}

\item{n}{The number of EOTs to show.}

\item{type}{Indicates which spatial representation should be shown. Either the regression weights as
in the original work of van Dool 2000 with \code{"regression weights"}
or squared correlation coefficient per grid cell with \code{"R2"}.}
}
\description{
This functions is a rather easy plotting function to see the spatial & temporal
representation of the first n calculated EOT modes.
Typically a maximum of 15 modes should not be exceeded.
}
